add: new segment for dynamic rendering as [word]
and updates on search component
