+++ src/components/customs/WordCard.tsx
@@ -1,62 +1,22 @@
-'use client'
-
 import React, {
   Fragment
 } from "react";
-import {
-  Card,
-  CardHeader,
-  CardBody,
-  CardFooter,
-  Divider,
-  Accordion,
-  AccordionItem,
-  Chip,
-} from "@nextui-org/react";
-import { Bookmark } from "lucide-react";
-import { api } from "@/src/trpc/react";
-import { toast } from "sonner";
-import { useLocale, useTranslations } from "next-intl";
+// import {
+//   Accordion,
+//   AccordionItem,
+// } from "@nextui-org/react";
+import { Card, CardHeader, CardBody, CardFooter } from "@nextui-org/card";
+import { Divider } from "@nextui-org/divider";
+import { Chip } from "@nextui-org/chip";
 import { WordSearchResult } from "@/types";
+import SaveWord from "./SaveWord";
 const itemClasses = {
   title: "font-normal text-fs-1 text-primary",
   trigger: "px-2 py-0 rounded-lg h-14 flex items-center",
   indicator: "text-fs-0",
   content: "px-2 text-fs--1",
 };
-export default function WordCard({ word: { word_data }, isSavedWord }: { word: WordSearchResult, isSavedWord?: boolean }) {
-  // const savedWords = api.user.getSavedWords.useQuery();
-
-  const locale = useLocale() as "en" | "tr";
-  const utils = api.useUtils()
-  const savedWordsQuery = api.user.getWordSaveStatus.useQuery(word_data.word_id, {
-    initialData: isSavedWord
-  })
-  const t = useTranslations("WordCard");
-  const saveWordMutation = api.user.saveWord.useMutation({
-    onMutate: async ({ wordId }) => {
-      await utils.user.getWordSaveStatus.cancel(wordId)
-      const previousValue = utils.user.getWordSaveStatus.getData(wordId);
-      utils.user.getWordSaveStatus.setData(wordId, !previousValue);
-      return { previousValue };
-    },
-    onError: (error, { wordId }, context) => {
-      switch (error.message) {
-        case "UNAUTHORIZED":
-          toast.error(t("UnauthSave"), {
-            position: "bottom-center",
-          });
-          utils.user.getWordSaveStatus.setData(wordId, context?.previousValue)
-          break;
-      }
-    },
-    // Always refetch after error or success:
-    onSettled: (newValue, error, { wordId }) => {
-      void utils.user.getWordSaveStatus.invalidate(wordId)
-    },
-  });
-
-
+export default function WordCard({ word: { word_data }, isSavedWord, locale }: { word: WordSearchResult, isSavedWord?: boolean, locale: "en" | "tr" }) {
   return (
     <Card
       aria-label="word card"
@@ -64,20 +24,7 @@ export default function WordCard({ word: { word_data }, isSavedWord }: { word: W
       isBlurred
       className="border border-border rounded-sm p-4"
     >
-      <button
-        className="absolute top-4 right-6 cursor-pointer z-50 sm:hover:scale-125 transition-all"
-        onClick={async () => {
-          await saveWordMutation.mutateAsync({ wordId: word_data.word_id });
-        }}
-        disabled={saveWordMutation.isPending}
-      >
-        <Bookmark
-          aria-label="bookmark icon"
-          size={32}
-          fill={savedWordsQuery.data ? "#F59E0B" : "#fff"}
-        />
-      </button>
-
+      <SaveWord word_data={word_data} isSavedWord={isSavedWord} />
       <CardHeader className="block sm:flex justify-center ">
         <h2 className="text-fs-4 w-full text-center sm:text-start">
           {word_data.prefix && (
